pyproject.toml
[project]
name = "untitled"
description = "a game for the PGC halloween jam 2023"
version = "0.1.0"
authors = [
    { name = "Unknown", email = "unknown@email.com" }
]
dependencies = [
    "pygame-ce >= 2.3.2",
    "pytmx >= 3.32"
]
requires-python = ">=3.11.0"


[tool.black]
line-length = 88

[tool.isort]
profile = "black"
